Separating out Apple Music notifications classes for readability. Fixing minor issue with broadcasting a complete message.
