+++ app/src/main/AndroidManifest.xml
@@ -110,7 +110,7 @@
             android:enabled="true"
             android:exported="true" />
         <service
-            android:name=".service.NotificationService"
+            android:name=".service.applemusic.NotificationService"
             android:label="AppleNotificationListener"
             android:permission="android.permission.BIND_NOTIFICATION_LISTENER_SERVICE"
             android:process=":remote"
