+++ app/src/main/java/com/adam/aslfms/service/applemusic/PlayingState.java
@@ -1,0 +1,12 @@
+package com.adam.aslfms.service.applemusic;
+
+/**
+ * Created by 4-Eyes on 16/3/2017.
+ *
+ */
+
+enum PlayingState {
+    UNKNOWN,
+    PLAYING,
+    PAUSED,
+}
